주문(고객) 추가

+++ src/main/java/shop/mtcoding/metamall/core/filter/JwtVerifyFilter.java
@@ -24,24 +24,24 @@ public class JwtVerifyFilter implements Filter {
         HttpServletRequest req = (HttpServletRequest) request;
         HttpServletResponse resp = (HttpServletResponse) response;
         String prefixJwt = req.getHeader(JwtProvider.HEADER);
-        if(prefixJwt == null){
+        if (prefixJwt == null) {
             error(resp, new Exception400("토큰이 전달되지 않았습니다"));
             return;
         }
         String jwt = prefixJwt.replace(JwtProvider.TOKEN_PREFIX, "");
         try {
             DecodedJWT decodedJWT = JwtProvider.verify(jwt);
-            int id = decodedJWT.getClaim("id").asInt();
+            Long id = decodedJWT.getClaim("id").asLong();
             String role = decodedJWT.getClaim("role").asString();
 
             // 세션을 사용하는 이유는 권한처리를 하기 위해서이다.
-            HttpSession session =  req.getSession();
+            HttpSession session = req.getSession();
             LoginUser loginUser = LoginUser.builder().id(id).role(role).build();
             session.setAttribute("loginUser", loginUser);
             chain.doFilter(req, resp);
-        }catch (SignatureVerificationException sve){
+        } catch (SignatureVerificationException sve) {
             error(resp, sve);
-        }catch (TokenExpiredException tee){
+        } catch (TokenExpiredException tee) {
             error(resp, tee);
         }
     }
